Backend/requirements.txt
# --- Deep Learning Core ---
torch
torchvision
numpy
ultralytics

# --- Image Processing ---
opencv-python-headless
pillow

# --- Malayalam Tools & NLP ---
symspellpy
transformers
sentencepiece

# --- Server & API ---
fastapi
uvicorn
python-multipart
pydantic

# --- Utilities ---
requests
gtts
reportlab
